Refactor to typescript
