+++ src/components/homepage/HomepageFeature.tsx
@@ -2,9 +2,7 @@ import React from 'react';
 import clsx from 'clsx';
 import styles from './HomepageFeature.module.css';
 
-
-
-export default function HomepageFeature( { id, Svg, img, title, description } ) {
+export default function HomepageFeature({ id, img, title, description }) {
   return (
     <div className={clsx('row', styles.feature, 'box')}>
       <div className={clsx('col col--6')}>
@@ -16,11 +14,9 @@ export default function HomepageFeature( { id, Svg, img, title, description } ) 
       <div className={clsx('col col--6', styles.content)}>
         <div className="text--center padding-horiz--md">
           <h2 className={styles.title}>{title}</h2>
-          <p  className={styles.subtitle}>{description}</p>
+          <p className={styles.subtitle}>{description}</p>
         </div>
       </div>
     </div>
   );
 }
-
-
