+++ src/components/homepage/HomepageFeatureAlt.tsx
@@ -2,15 +2,13 @@ import React from 'react';
 import clsx from 'clsx';
 import styles from './HomepageFeatureAlt.module.css';
 
-
-
-export default function HomepageFeatureAlt( { id, Svg, img, title, description } ) {
+export default function HomepageFeatureAlt({ id, img, title, description }) {
   return (
     <div className={clsx('row', styles.feature, 'box')}>
       <div className={clsx('col col--6', styles.content)}>
         <div className="text--center padding-horiz--md">
           <h2 className={styles.title}>{title}</h2>
-          <p  className={styles.subtitle}>{description}</p>
+          <p className={styles.subtitle}>{description}</p>
         </div>
       </div>
       <div className={clsx('col col--6')}>
@@ -22,5 +20,3 @@ export default function HomepageFeatureAlt( { id, Svg, img, title, description }
     </div>
   );
 }
-
-
